requirements.txt
Flask==3.1.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
Flask-Mail==0.9.1
flask-cors==6.0.2
Flask-Caching==2.3.1
WTForms==3.0.1
SQLAlchemy==2.0.49
python-dotenv==1.0.0
email-validator==2.1.0
marshmallow==4.3.0
pandas==3.0.2
openpyxl==3.1.2
gunicorn==21.2.0
resend==2.27.0
